// fe/src/features/rooms/types/rooms.ts
export type Location = {
  id: number
  tenViTri: string
  tinhThanh: string
  quocGia: string
  hinhAnh: string
}

export type Room = {
  id: number
  tenPhong: string
  khach: number
  phongNgu: number
  giuong: number
  phongTam: number
  moTa: string
  giaTien: number
  mayGiat: boolean
  banLa: boolean
  tivi: boolean
  dieuHoa: boolean
  wifi: boolean
  bep: boolean
  doXe: boolean
  hoBoi: boolean
  banUi: boolean
  maViTri: number
  hinhAnh: string
  createdAt: string
  updatedAt: string
  viTri: Location
}

export type RoomsResponse = {
  message: string
  data: Room[]
}